+++ python/tests/collections/volume_test.py
@@ -129,7 +129,7 @@ def test_volume_minimum_attribs(anthology):
     )
     assert volume.full_id == "L05-6"
     assert volume.title == "Lorem ipsum"
-    assert volume.get_ingest_date().year == 1900
+    assert volume.ingest_date.year == 1900
     assert not volume.is_workshop
 
 
@@ -152,16 +152,14 @@ def test_volume_all_attribs(anthology):
         shortbooktitle="L.I.",
         venue_ids=["li", "acl"],
     )
-    assert volume.ingest_date == "2023-01-12"
-    assert volume.get_ingest_date() == date(2023, 1, 12)
+    assert volume.ingest_date == date(2023, 1, 12)
 
 
 def test_volume_attributes_2022acl_long(anthology):
     volume = anthology.get_volume("2022.acl-long")
     assert isinstance(volume, Volume)
     assert volume.id == "long"
-    assert volume.ingest_date == "2022-05-15"
-    assert volume.get_ingest_date() == date(2022, 5, 15)
+    assert volume.ingest_date == date(2022, 5, 15)
     assert volume.address == "Dublin, Ireland"
     assert volume.publisher == "Association for Computational Linguistics"
     assert volume.doi is None
@@ -179,8 +177,7 @@ def test_volume_attributes_2022acl_demo(anthology):
     volume = anthology.get_volume("2022.acl-demo")
     assert isinstance(volume, Volume)
     assert volume.id == "demo"
-    assert volume.ingest_date == "2022-05-15"
-    assert volume.get_ingest_date() == date(2022, 5, 15)
+    assert volume.ingest_date == date(2022, 5, 15)
     assert volume.address == "Dublin, Ireland"
     assert volume.publisher == "Association for Computational Linguistics"
     assert volume.doi == "10.18653/v1/2022.acl-demo"
@@ -231,10 +228,9 @@ def test_volume_without_frontmatter(anthology):
 def test_volume_set_ingest_date(anthology):
     volume = anthology.get_volume("2022.acl-demo")
     volume.ingest_date = "2025-07-15"
-    assert volume.get_ingest_date() == date(2025, 7, 15)
+    assert volume.ingest_date == date(2025, 7, 15)
     volume.ingest_date = date(2026, 3, 1)
-    assert volume.get_ingest_date() == date(2026, 3, 1)
-    assert volume.ingest_date == "2026-03-01"
+    assert volume.ingest_date == date(2026, 3, 1)
 
 
 @pytest.mark.parametrize(
@@ -449,7 +445,7 @@ def test_volume_create_paper_implicit(anthology):
     assert volume.collection.is_modified
     assert paper.authors == authors
     assert paper.title.as_text() == "The awesome paper I have never written"
-    assert paper.ingest_date == "2025-01-07"
+    assert paper.ingest_date.isoformat() == "2025-01-07"
     assert paper.parent is volume
     assert paper.id in volume
     # Highest paper ID in 2022.acl-long is 603, so this one should automatically get 604
@@ -473,7 +469,7 @@ def test_volume_create_paper_explicit(anthology):
     assert volume.collection.is_modified
     assert paper.authors == authors
     assert paper.title.as_text() == "The awesome paper I have never written"
-    assert paper.ingest_date == "2025-01-07"
+    assert paper.ingest_date.isoformat() == "2025-01-07"
     assert paper.parent is volume
     assert paper.id in volume
     assert paper.id == "701"