+++ python/CHANGELOG.md
@@ -23,6 +23,7 @@
 - `Venue.item_ids` will now update dynamically when a Volume's `venue_ids` attribute is modified, or when a new Volume is created.
 - `Paper.get_issue()` is deprecated in favor of `Paper.journal_issue`.
 - `{Paper|Volume}.get_journal_title()` is deprecated in favor of `.journal_title`.
+- `{Paper|Volume}.get_ingest_date()` is deprecated in favor of `.ingest_date`, which now always returns a `date` object instead of a string.
 
 ### Removed